requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.52.0
pyyaml==6.0.2
